revert commands refactor
