--- core/src/main/scala/org/apache/spark/sql/crossdata/catalyst/execution/DoCatalogDataSourceTable.scala
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) 2015 Stratio (http://stratio.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.spark.sql.crossdata.catalyst.execution
-
-import org.apache.spark.sql.crossdata.XDContext
-import org.apache.spark.sql.crossdata.catalog.XDCatalog.CrossdataTable
-import org.apache.spark.sql.execution.RunnableCommand
-import org.apache.spark.sql.{Row, SQLContext}
-
-private[crossdata] trait DoCatalogDataSourceTable extends RunnableCommand {
-
-  protected val crossdataTable: CrossdataTable
-  protected val allowExisting: Boolean
-
-  override def run(sqlContext: SQLContext): Seq[Row] =
-    catalogDataSourceTable(sqlContext.asInstanceOf[XDContext])
-
-  protected def catalogDataSourceTable(crossdataContext: XDContext): Seq[Row]
-
-}
