+++ core/src/main/scala/org/apache/spark/sql/crossdata/catalyst/execution/commands.scala
@@ -1,34 +1,52 @@
-/*
- * Copyright (C) 2015 Stratio (http://stratio.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package org.apache.spark.sql.crossdata.catalyst.execution
 
-import org.apache.spark.sql._
 import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.catalyst.analysis.EliminateSubQueries
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.crossdata.XDContext
+import org.apache.spark.sql._
 import org.apache.spark.sql.crossdata.catalog.XDCatalog.CrossdataTable
 import org.apache.spark.sql.crossdata.catalog.interfaces.XDCatalogCommon
+import org.apache.spark.sql.crossdata.util.CreateRelationUtil._
 import org.apache.spark.sql.execution.RunnableCommand
 import org.apache.spark.sql.execution.datasources.{LogicalRelation, ResolvedDataSource}
 import org.apache.spark.sql.sources.{HadoopFsRelation, InsertableRelation}
 import org.apache.spark.sql.types.StructType
 
-private[crossdata]
-case class PersistSelectAsTable(
+private[crossdata] trait DoCatalogDataSourceTable extends RunnableCommand {
+
+  protected val crossdataTable: CrossdataTable
+  protected val allowExisting: Boolean
+
+  override def run(sqlContext: SQLContext): Seq[Row] =
+    catalogDataSourceTable(sqlContext.asInstanceOf[XDContext])
+
+  protected def catalogDataSourceTable(crossdataContext: XDContext): Seq[Row]
+
+}
+
+private[crossdata] case class PersistDataSourceTable(
+                                                      protected val crossdataTable: CrossdataTable,
+                                                      protected val allowExisting: Boolean
+                                                    ) extends DoCatalogDataSourceTable {
+
+  override protected def catalogDataSourceTable(crossdataContext: XDContext): Seq[Row] = {
+
+
+    val tableIdentifier = crossdataTable.tableIdentifier
+
+    if (crossdataContext.catalog.tableExists(tableIdentifier.toTableIdentifier) && !allowExisting)
+      throw new AnalysisException(s"Table ${tableIdentifier.unquotedString} already exists")
+    else
+      crossdataContext.catalog.persistTable(crossdataTable, createLogicalRelation(crossdataContext, crossdataTable))
+
+    Seq.empty[Row]
+  }
+
+
+}
+
+private[crossdata] case class PersistSelectAsTable(
                                  tableIdent: TableIdentifier,
                                  provider: String,
                                  partitionColumns: Array[String],
@@ -124,3 +142,24 @@ case class PersistSelectAsTable(
     }
   }
 }
+
+
+private[crossdata] case class RegisterDataSourceTable(
+                                                       protected val crossdataTable: CrossdataTable,
+                                                       protected val allowExisting: Boolean
+                                                     ) extends DoCatalogDataSourceTable {
+
+  override protected def catalogDataSourceTable(crossdataContext: XDContext): Seq[Row] = {
+
+    val tableIdentifier = crossdataTable.tableIdentifier.toTableIdentifier
+
+    crossdataContext.catalog.registerTable(
+      tableIdentifier,
+      createLogicalRelation(crossdataContext, crossdataTable),
+      Some(crossdataTable)
+    )
+    Seq.empty[Row]
+  }
+
+}
+
